java/refactoring/src/main/java/ru/akirakozov/sd/refactoring/web/servlet/AddProductServlet.java
package ru.akirakozov.sd.refactoring.web.servlet;

import ru.akirakozov.sd.refactoring.model.domain.ProductUtils;
import ru.akirakozov.sd.refactoring.model.repository.ProductRepository;

import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

/**
 * @author akirakozov
 */
public class AddProductServlet extends HttpServlet {
    private final ProductRepository productRepository;

    public AddProductServlet(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    @Override
    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
        String name = request.getParameter("name");
        long price = Long.parseLong(request.getParameter("price"));
        productRepository.save(ProductUtils.makeProduct(name, price));

        response.setContentType("text/html");
        response.setStatus(HttpServletResponse.SC_OK);
        response.getWriter().println("OK");
    }
}
